Ticket 25 fixes

+++ routes.js
@@ -83,6 +83,10 @@ router.get("/getBand/:id", async (req, res) => {
     res.json(await dbconnection.getBand(req.params.id));
 })
 
+router.get("/getTakenBandLevels", async (req, res) => {
+    res.json(await dbconnection.getTakenBandLevels());
+})
+
 router.get("/getAssociatedTrainingIDsWithBand/:id", async (req, res) => {
     res.json(await dbconnection.getAssociatedTrainingIDsWithBand(req.params.id));
 })
@@ -144,12 +148,15 @@ router.post("/deleteJobFamily", async (req, res) => {
 
 router.post("/deleteBand", async (req, res) => {
     //delete links in junction table associated with band
-    let result = await dbconnection.deleteAssociatedTrainingsWithBand(req.body.BandID);
-    result = await dbconnection.deleteAssociatedCompetenciesWithBand(req.body.BandID);
-    result = await dbconnection.deleteBand(req.body.BandID);
-  // #swagger.description = 'deletes an existing band by BandID'
-    res.json(result);
-
+    if (!await dbconnection.canDeleteBand(req.body.BandID)) {
+        res.json("error");
+    } else {
+        let result = await dbconnection.deleteAssociatedTrainingsWithBand(req.body.BandID);
+        result = await dbconnection.deleteAssociatedCompetenciesWithBand(req.body.BandID);
+        result = await dbconnection.deleteBand(req.body.BandID);
+      // #swagger.description = 'deletes an existing band by BandID'
+        res.json(result);
+    }
 })
